Move twitter comps to their owwn folder
