+++ twindle-cli/src/renderer/main/twitter/components/App.svelte
